Use WebACRolesProvider to check ACLs in WebACAuthorizationRealm
Inject the servlet request into the WebACAuthorizationRealm so that it can find the node (or its parent, if the node doesn't exist yet) corresponding to the request URI. Once it has obtained that node, calls the WebACRolesProvider to get a list of user => mode mappings for the resource. These are translated into WebACPermission objects and added to the authorization info.

Additional fixes and cleanup:
* Removed the FAD and authN provider configuration from the repo.xml Spring configuration. The only remaining WebAC bean in repo.xml is the WebACRolesProvider, since it expects and injected NodeService.
* Fixed up the TestAuthenticationRequestFilter to properly return null for getUserPrincipal() when the request was anonymous.
* Fixed up the ServletContainerAuthenticatingRealm to use the correct constructor for the SimpleAuthenticationInfo class.
* Ensure anonymous users are not authenticated in Shiro by calling currentUser.logout() when an anonymous request is detected.
* Run permission check on anonymous users.

Fixes: https://jira.duraspace.org/browse/FCREPO-2760